Update and rename libirecovery-0.1.1_p20180328.ebuild to libirecovery-0.1.1_p20180917.ebuild
libirecovery 0.1.1_p20180328 -> 0.1.1_p20180917
Much overdue version bump
Synced with latest upstream commit
Added support for new --with-dummy configure option
This new option means that libirecovery doesn't require USB support if it is activated.
This means that libirecovery effectively only is useful as a query tool, if this option is on.
Also, this is the first ~amd64 version of the ebuild. I consider this tool more or less stable enough
to merit an ~arch KEYWORD on the arch I _have_ tested it on (amd64).